% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.r
\name{abund_env_grad}
\alias{abund_env_grad}
\title{Environmental gradient and Species abundances}
\usage{
abund_env_grad(
  know_env_grad_pos = TRUE,
  nb_resource = 40,
  nb_consumer = 100,
  nb_location = 3,
  mean_tol_env = 1,
  sd_tol_env = 1,
  env_grad_min = 0,
  env_grad_max = 10
)
}
\arguments{
\item{know_env_grad_pos}{Boolean, If True: then evenly spaced sampling along the environmental
gradient, otherwise, uniformly drawn position along the gradient}

\item{nb_resource}{Numeric, Number of resources in the bipartite network}

\item{nb_consumer}{Numeric, Number of consumer in the bipartite network}

\item{nb_location}{Numeric, Number of location along the gradient (i.e. number of networks)}

\item{mean_tol_env}{Numeric, Mean of the standard deviation of the niche tolerance}

\item{sd_tol_env}{Numeric, Standard deviation of the standard deviation of the niche tolerance}

\item{env_grad_min}{Numeric, Maximum value of the gradient in arbitrary unit}

\item{env_grad_max}{Numeric, Minimum value of the gradient in arbitrary unit}
}
\value{
List, Two arrays in a list containing respectively the resource and consumer abundance along the gradient
}
\description{
Generate the environmental gradient and abundances for the species along this gradient,
separately for the consumer and the resources.
}
\examples{
abund_env_grad(nb_resource = 40, nb_consumer = 100, nb_location = 3,
               mean_tol_env = 1, sd_tol_env = 1,
               env_grad_min = 0, env_grad_max = 10)
}
