% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.r
\name{env_grad_netw}
\alias{env_grad_netw}
\title{Network sampling generation across an environmental gradient}
\usage{
env_grad_netw(
  nb_resource = 40,
  nb_consumer = 100,
  nb_location = 3,
  know_env_grad_pos = TRUE,
  mean_tol_env = 1,
  sd_tol_env = 1,
  env_grad_min = 0,
  env_grad_max = 10,
  le_grad = 100,
  ratio_grad = 0.8,
  buffer = 1,
  mean_tol = 2,
  sd_tol = 10,
  delta = 1,
  ninter = 100
)
}
\arguments{
\item{nb_resource}{Numeric, Number of resources in the bipartite network}

\item{nb_consumer}{Numeric, Number of consumers in the bipartite network}

\item{nb_location}{Numeric, Number of location along the gradient (i.e. number of networks)}

\item{know_env_grad_pos}{Boolean, If True: then evenly spaced sampling along the environmental
gradient, otherwise, uniformly drawn position along the gradient}

\item{mean_tol_env}{Numeric, Mean of the standard deviation of the niche tolerance}

\item{sd_tol_env}{Numeric, Mean of the standard deviation of the niche tolerance}

\item{env_grad_min}{Numeric, Maximum value of the gradient in arbitrary unit}

\item{env_grad_max}{Numeric, Minimum value of the gradient in arbitrary unit}

\item{le_grad}{Numeric, Higher bound of the first trait gradient}

\item{ratio_grad}{Numeric between 0 and 1, Ratio between the first and second gradient}

\item{buffer}{Numeric, Enable the mean of the trait to fall outside of the gradient by the buffer value}

\item{mean_tol}{Numeric, Mean of the standard deviation of the trait tolerance}

\item{sd_tol}{Numeric, Standard deviation of the standard deviation of the trait tolerance}

\item{delta}{Numeric, weight of trait matching relatively to abundance}

\item{ninter}{Numeric, number of interactions observed on the network}
}
\value{
List of matrix, Observed interactions across an environmental gradient
}
\description{
Generate and environmental gradient and random trait matching to create an theoretical interaction
probability and then sample this theoretical network to simulate random observation.
}
