% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.r
\name{sampling}
\alias{sampling}
\title{Sampling with Multinomial}
\usage{
sampling(th_network, ninter = 100, nb_resource = nb_resource)
}
\arguments{
\item{th_network}{Matrix, Theoretical interaction probability taking in account trait matching and abundances}

\item{ninter}{Numeric, number of interactions observed on the network}

\item{nb_resource}{Numeric, Number of resources in the bipartite network}
}
\value{
Matrix, Observed interactions across an environmental gradient
}
\description{
Using the multinomial law we can sample the theoretical network according to the probability of two
species interaction's probability.
}
