% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.r
\name{trait_match_mat}
\alias{trait_match_mat}
\title{Trait matching matrix}
\usage{
trait_match_mat(
  ratio_grad = 0.8,
  mean_tol = 0.1,
  sd_tol = 1,
  buffer = 1,
  nb_resource = 40,
  nb_consumer = 100,
  corr_env_trait = 0,
  param_epsilon = c(0.5, 0.1),
  trait_env_res,
  trait_env_con
)
}
\arguments{
\item{ratio_grad}{Numeric between 0 and 1, Ratio between the first and second gradient}

\item{mean_tol}{Numeric, Mean of the standard deviation of the trait tolerance}

\item{sd_tol}{Numeric, Standard deviation of the standard deviation of the trait tolerance}

\item{buffer}{Numeric, Enable the mean of the trait to fall outside of the gradient by the buffer value}

\item{nb_resource}{Numeric, Number of resource in the bipartite network}

\item{nb_consumer}{Numeric, Number of consumer in the bipartite network}

\item{corr_env_trait}{Numeric, weight given on the correlation between the environment gradient and the trait one}

\item{param_epsilon}{Vector of two values, containing the mean and sd of the epsilon of correlation of the first trait}

\item{trait_env_res}{Matrix, obtained from abund_env_grad containing the th traits of the resource}

\item{trait_env_con}{Matrix, obtained from abund_env_grad containing the th traits of the consumer}
}
\value{
List, Array containing the probabilities of species interactions only based on trait matching
and then 2 Arrays containing the trait distribution of both traits respectively
}
\description{
Draw 2 random traits with different optimums and variance for each species and then induce a trait
matching probability for each couple of species in a Matrix.
}
