% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.r
\name{env_grad_netw}
\alias{env_grad_netw}
\title{Network sampling generation across an environmental gradient}
\usage{
env_grad_netw(
  nb_resource = 40,
  nb_consumer = 100,
  nb_location = 3,
  know_env_grad_pos = TRUE,
  mean_tol_env = 0.1,
  sd_tol_env = 0.1,
  magn_res_min = 20,
  magn_res_max = 200,
  magn_con_min = 100,
  magn_con_max = 10000,
  corr_env_trait = 0,
  param_epsilon = c(0.5, 0.1),
  ratio_grad = 0.8,
  buffer = 1,
  mean_tol = 2,
  sd_tol = 10,
  delta = 1,
  ninter = 100
)
}
\arguments{
\item{nb_resource}{Numeric, Number of resources in the bipartite network}

\item{nb_consumer}{Numeric, Number of consumers in the bipartite network}

\item{nb_location}{Numeric, Number of location along the gradient (i.e. number of networks)}

\item{know_env_grad_pos}{Boolean, If True: then evenly spaced sampling along the environmental
gradient, otherwise, uniformly drawn position along the gradient}

\item{mean_tol_env}{Numeric, Mean of the standard deviation of the niche tolerance for an environmental gradient between 0 and 1}

\item{sd_tol_env}{Numeric, Standard deviation of the standard deviation of the niche tolerance for an environmental gradient between 0 and 1}

\item{magn_res_min}{Numeric, Lower bound of the order of magnitude of size of a resource population}

\item{magn_res_max}{Numeric, Higher bound of the order of magnitude of size of a resource population}

\item{magn_con_min}{Numeric, Lower bound of the order of magnitude of size of a consumer population}

\item{magn_con_max}{Numeric, Higher bound of the order of magnitude of size of a consumer population}

\item{corr_env_trait}{Numeric, weight given on the correlation between the environment gradient and the trait one}

\item{param_epsilon}{Vector of two values, containing the mean and sd of the epsilon of correlation of the first trait}

\item{ratio_grad}{Numeric between 0 and 1, Ratio between the first and second gradient}

\item{buffer}{Numeric, Enable the mean of the trait to fall outside of the gradient by the buffer value}

\item{mean_tol}{Numeric, Mean of the standard deviation of the trait tolerance}

\item{sd_tol}{Numeric, Standard deviation of the standard deviation of the trait tolerance}

\item{delta}{Numeric, weight of trait matching relatively to abundance}

\item{ninter}{Numeric, number of interactions observed on the network}
}
\value{
List of matrix, Observed interactions across an environmental gradient,
as well as the abundances and related distribution information,
the theoretical network,
the trait distribution and trait matching probabilities.
}
\description{
Generate and environmental gradient and random trait matching to create an theoretical interaction
probability and then sample this theoretical network to simulate random observation.
}
