% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.r
\name{int_count_th}
\alias{int_count_th}
\title{Mix abundance and trait matching}
\usage{
int_count_th(location, abund_resource, abund_consumer, p_matching, delta = 1)
}
\arguments{
\item{location}{Numeric, Number of location along the gradient (i.e. number of networks)}

\item{abund_resource}{Array, resource abundance along the gradient}

\item{abund_consumer}{Array, consumer abundance along the gradient}

\item{p_matching}{Matrix, Array containing the probabilities of species interactions only based on trait matching}

\item{delta}{Numeric, weight of trait matching relatively to abundance}
}
\value{
Matrix, Theoretical interaction probability taking in account trait matching and abundances
}
\description{
Take in account the probability of matching only by meeting randomly (with the abundances) and the
of interacting due to trait matching.
}
