% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.r
\name{trait_match_mat}
\alias{trait_match_mat}
\title{Trait matching matrix}
\usage{
trait_match_mat(
  le_grad = 100,
  ratio_grad = 0.8,
  mean_tol = 2,
  sd_tol = 10,
  buffer = 1,
  nb_resource = 40,
  nb_consumer = 100
)
}
\arguments{
\item{le_grad}{Numeric, Higher bound of the first trait gradient}

\item{ratio_grad}{Numeric between 0 and 1, Ratio between the first and second gradient}

\item{mean_tol}{Numeric, Mean of the standard deviation of the trait tolerance}

\item{sd_tol}{Numeric, Standard deviation of the standard deviation of the trait tolerance}

\item{buffer}{Numeric, Enable the mean of the trait to fall outside of the gradient by the buffer value}

\item{nb_resource}{Numeric, Number of resource in the bipartite network}

\item{nb_consumer}{Numeric, Number of consumer in the bipartite network}
}
\value{
Matrix, Array containing the probabilities of species interactions only based on trait matching
}
\description{
Draw 2 random traits with different optimums and variance for each species and then induce a trait
matching probability for each couple of species in a Matrix.
}
